Updated MembersFeature to allow Admins to promote or revoke user as club member (this is added to  the New and edit user section also in delete Users)

+++ App/Http/Controllers/Admin/UsersController.php
@@ -9,6 +9,9 @@ use App\Http\Functions\Validate;
 use App\Http\Libraries\Authentication\Auth;
 use App\Http\Libraries\Authentication\Authenticate;
 use App\Http\Libraries\Authentication\Csrf;
+use App\Http\Models\FeaturedImage;
+use App\Http\Models\Image;
+use App\Http\Models\Member;
 use App\Http\Models\Profile;
 use App\Http\Models\User;
 use App\Http\Models\UserSettings;
@@ -58,7 +61,7 @@ class UsersController
                         $users = new User();
                         $users->username = $validate->Post("username");
                         $users->email = $validate->Post("email");
-                        $users->username = password_hash($validate->Post("password"),PASSWORD_DEFAULT);
+                        $users->password = password_hash($validate->Post("password"), PASSWORD_DEFAULT);
                         $users->save();
 
                         $user->username = $validate->Post("username");
@@ -76,6 +79,14 @@ class UsersController
                         $settings->display_dob = 1;
                         $settings->display_email = 1;
                         $settings->save();
+
+                        if ($validate->Post("make_member") == 1) {
+//                    Will check if the member does not exist and will create a new one;
+                            $member = new Member();
+                            $member->user_id = $users->id;
+                            $member->save();
+                        } else {
+                        }
 
                         redirect($url->make("admin.users.home"));
                     }
@@ -102,29 +113,50 @@ class UsersController
         $id = base64_decode($id);
         $username = base64_decode($username);
         $user = User::withCount("settings")->where("id", $id)->where("username", $username)->get();
+        $members = $user->first()->Members()->where("user_id", $user->first()->id)->get();
+        echo $members;
         if ($user->count() == 1) {
-            echo TemplateEngine::View("Pages.Admin.Users.edit", ["user" => $user->first(), "url" => $url]);
+            echo TemplateEngine::View("Pages.Admin.Users.edit", ["user" => $user->first(), "url" => $url, "members" => $members]);
         } else {
-            echo "Article doesnt exisit";
+            echo "user} doesnt exisit";
         }
 
 
     }
 
-    public function update(Url $url, Csrf $csrf)
+    public function update(Url $url, Csrf $csrf, Validate $validate)
     {
 //Get validation
 
         if ($csrf->Verify() == true) {
             $validate = new Validate();
+            $id = $validate->Post("id");
             $user = User::find($validate->Post("id"));
             $user->email = $validate->Required("email")->Post();
+            $user->is_admin = $validate->Post("is_admin");
             $user->save();
 //
             $profile = Profile::find($user->id);
             $profile->first_name = $validate->Required("first_name")->Post();
             $profile->last_name = $validate->Required("last_name")->Post();
             $profile->save();
+
+
+            if ($validate->Post("make_member") == 1) {
+                $member = Member::where("user_id", $id)->get();
+                if ($member->count() == 0) {
+//                    Will check if the member does not exist and will create a new one;
+                    $member = new Member();
+                } else {
+                    $member = $member->first();
+                }
+                $member->user_id = $id;
+                $member->save();
+            } else {
+                $member = Member::where("user_id", $validate->Post("id"));
+                $member->delete();
+            }
+
             redirect($url->make("admin.users.home"));
         }
 
@@ -132,9 +164,31 @@ class UsersController
 
     public function delete($id, Url $url)
     {
-        $user = User::find($id);
-        $user->delete();
+        $user = User::where("id",$id);
+        if($user->count() == 1) {
+            UserSettings::where("user_id", $id)->delete();
+            Profile::where("user_id", $id)->delete();
+            Member::where("user_id", $id)->delete();
+
+
+//        FInd Images and delete them
+            $images = Image::where("user_id", $id);
+            if($images->count() >= 1) {
+                foreach ($images as $image) {
+                    unlink(__DIR__ . "/../../../../img/uploads/$image->image_name");
+                    Image::find($image->id)->delete();
+                    FeaturedImage::where("image_id", $image_id)->delete();
+                }
+            }
+//            finally delete the users account;
+            $user->delete();
+        }
+        else
+        {
+            echo "no user found";
+        }
         redirect($url->make("admin.users.home"));
+
     }
